network-emulation: Allow fine grained packet loss
Before this change, packet loss was specified in an integral number of
percentage. A packet loss of one percent is quite much, and as the
underlying model used packet loss as a double between [0.0 and 1.0],
this is now exposed in the API.

The previous loss_percent configuration is still available, but marked
as deprecated.

Bug: webrtc:13344
Change-Id: I9dfc602619fb962572b640e05f5f7e811da4d3d2