Introduce RTC_NO_DESTROY annotation.
This macro introduces an annotation to tell the compiler that a variable
with static or thread storage duration shouldn't have its
exit-time destructor run.

Apply this to code that is currently compiled using the
-Wno-exit-time-destructors flag.

LSAN will be unhappy with this annotation, but WebRTC itself hasn't been
tested with LSAN enabled, and it shouldn't be a problem as all code
trying to apply this does not build with build_with_chromium=true.

Bug: webrtc:9693, webrtc:11665
Change-Id: Ia28e3534170e0817b815717f6efe862f7b51ef62