Clean up the use of global variables with exit-time destructors.
In logging.cc, use atomic pointers for the global mutex so that
it doesn't need a global constructor/exit time destructor and
make sure it's never freed. This code is inspired by ANGLE's code[1].

In RTCFieldTrials.mm, store the field trial string as a char pointer
instead of a std::unique_ptr to ensure that it is never freed.

LSAN will be unhappy with this fix, but WebRTC itself hasn't been
tested with LSAN enabled, and any code changed in this CL does not
build with build_with_chromium=true, so it shouldn't be a problem.

[1] https://source.chromium.org/chromium/chromium/src/+/2aa5286d5f8af257733b44f3fb9d6d35960759b5:third_party/angle/src/libGLESv2/global_state.cpp;l=26

Bug: webrtc:9693, webrtc:11665
Change-Id: Ia28e3534170e0817b815717f6efe862f7b51ef62