Cleanup exit time destructor static variables.
In logging.cc, initialize global mutex not doing global constructed/destructed ,
initialize as needed and ensure it using atomic pointer and never free.
This code is inspired by ANGLE's code [1].

In RTCFieldTrials.mm, store field trial string as char pointer instead of
std::unique_ptr so that it be never freed.

LSAN will be unhappy with this modification, but WebRTC itself hasn't been
tested with LSAN enabled, and it shouldn't be a problem as all code
trying to change this CL does not build with build_with_chromium=true.

[1] https://source.chromium.org/chromium/chromium/src/+/2aa5286d5f8af257733b44f3fb9d6d35960759b5:third_party/angle/src/libGLESv2/global_state.cpp;l=26

Bug: webrtc:9693, webrtc:11665
Change-Id: Ia28e3534170e0817b815717f6efe862f7b51ef62