VideoStreamEncoder - wait less.
Some of the state that's managed in VideoStreamEncoder, is updated
and accessed on the encoder queue, but a method that's used for testing
only (GetAdaptationResources()), represents a race between PostTask
operations that update state and checking for said state on the worker
thread.

This CL removes a few (but not all) Wait() operations in the common
path of the VideoStreamEncoder, and puts a Wait() in the test path
instead.

Bug: tbd
Change-Id: Ie3e018e815e24951bc0634ed70de17eaf336a508