remotedesktop_portal: Add mechanism to retreive session related information
Remotedesktop portal is capable of both capturing screen and controlling
the input devices. However, since remote desktop portal is instantiated
under the desktop capture, we woulnd't want to overload the impementation
here to control devices and would like to keep the logic to inject input
into separate input injector module (one already exists in chromoting code
base for X11, for example). This change exposes an API so that the remote
desktop session information can be obtained from the remote desktop
capturer and reused in the wayland input injector to inject input.

Bug: chromium:1291247
Change-Id: Ie74f326db89ba1b7680f5beaeae6718bb62e4cf6