Fix fd leak in ifaddrs_android.cc and introduce rtc_base/scoped_generic.h.
If `populate_ifaddrs` in `getifaddrs` returns a non-zero value,
the fd opened in `getifaddrs` will not be closed.
To prevent this mistake, add scoepd_generic.h copied from
Chromium's base/scoped_generic.h and apply it.
The copied code is almost identical to Chromium's code,
except for removing the use of raw_ptr.

Bug: webrtc:13674, webrtc:13082
Change-Id: I1eb01c1c68d9486ee5402d33963b86d18dfd84b1