Ensure PipeWire doesn't use a Null SourceId
This has mostly seemed to work fine until now; but there's a collision
happening in chromium where if the source is being shown in the Window
Picker it collides with the (also null) Dialog ID and is ignored. While
we could patch that code to not count Null as a collision, there's the
potential for other (future) code to simply ignore a capture source
that it thinks is Null.

Fixed: chromium:1295375
Change-Id: I4356084f0af97f4d56632938b0d9a24d327f7107