[DesktopCapture][WGC] Avoid artifacts when capture source is resized
This CL fixes the issue where artifacts appear during capture with WGC
when the capture source is resized. A video of the issue is available
here: https://bugs.chromium.org/p/webrtc/issues/detail?id=9273#c44

The solution is to use CopySubresourceRegion instead of CopyResource to
only copy valid data into our texture. Additionally, we moved the call
to CreateMappedTexture to before the call to CopySubresourceRegion, as
the latter requires both textures to be of the same size.

Bug: webrtc:9273
Change-Id: I114458d95cbf58550ff653a985dd84db4741e0f8