WebRTC-DeprecateGlobalFieldTrialString/Enabled/ - part 6/inf
This patch just refactors creation of P2P transport channel,
pushing down the IceTransportInit object rather than decomposing
it going down.

The IceTransportInit object will in subsequent patches be
extended with a field trial container.

Reason for splitting patch into this and subsequent is
to allow changes to internal factories.

Bug: webrtc:10335
Change-Id: Icc8b6e4142744b64d134bcb2d4a56777745db62b