Detect DXGI capturer hangs when changing resolution in detached sessions
This CL addresses an issue where the desktop appears to freeze after
resizing the desktop in a curtained CRD session when using the DXGI
capturer. This problem does not reproduce when using the GDI capturer nor
does it reproduce when the Windows session is attached to the local
console.

After digging in, it appears that the DXGI DuplicateOutput API stops
providing updated frame data. No errors are returned but yet no data is
produced. The problem is that when in this condition, there isn't a good
way to discern between this problem and a case where the desktop is
actually static.

The DxgiDuplicatorController already contains logic to attempt to capture
a frame prior returning success afte reinitialization.  This logic works
fine in the console case and occasionally works in the detached session
case. What I noticed in my reproductions was that DXGI would produce a few
frames before hanging (usually 1-2 but sometimes 3 or 4).  My solution is
to check the session state and adjust the number of frames we attempt to
capture (I also simplified the wait logic as there was a bug in the time
calc and it seemed more complicated than it needed to be).

One option considered would be to introduced a new differ class higher up
in the stack which would run the GDI and DXGI capturers in parallel
(instead of in the fallback configuration as they are today) however that
seemed like overkill for this specific issue.

Bug: chromium:1307357
Change-Id: Idba4bb9b2aa7692040344d480be3f0d09b9ce9e9