Implment IsAdapterAvailable in AndroidNetworkMonitor
This cl/ fixes a race condition with the recent additions
to NetworkMonitorAutoDetect (getAllNetworksFromCache).

The getAllNetworksFromCache-feature uses the by the Android team preferred way of
enumarating networks, i.e to register network listeners.

This however introduces a unpleasant race condition like this:
1) network.cc discover rmnet0
2) BasicPortAllocator tries to create UDP port on rmnet0
- This fails as BindSocketToNetwork requires a android handle.
3) NetworkMonitorAutoDetect gets callback with rmnet0
4) BasicPortAllocator tries to create TCP port on rmnet0
- This succeeds.
5) Since rmnet0 has one working port, there will not be
any new ports created on that network

=> We end up with a TCP only connection :(

---

By impl. IsAdapterAvaiable, we make sure that the network
will not be used by BasicPortAllocator (or anyone else!)
until we support binding to it.

This cl/ is default enabled with the kill-switch
WebRTC-AndroidNetworkMonitor-IsAdapterAvailable.

Bug: webrtc:13741
Change-Id: I7c2cb7789660fd2e082c214d00e50c894666b07c