Adding some AV1 constants and helper functions
I've added a basic AV1 impl to Chrome Remote Desktop and am looking into
what is needed to test with I444 (Profile-1) in our platform. This CL
adds a few helper functions, constants, and enums that can be used to
configure the SDP with different AV1 profiles. More work is still needed
but I wanted to get this in place first so I can build on it in the CRD
host code.

Bug: chromium:1329660

Change-Id: I1af9ebf31f833138e8c36e0c0a30e32289e7b58e