RtcEventLogImpl: avoid most PostTasks.
Research in https://webrtc-review.googlesource.com/c/src/+/250165
shows that RtcEventLogImpl::Log PostTask traffic can be removed
almost entirely by pretending to be the destination queue and
executing the lambda inline. This change goes further and
additionally avoids memory allocation for the lambda if the inline
version of PostTask is used.

Bug: None so far
Change-Id: I940f058fe5badecfd1685832da81a7079c28c0a9