When updating the max loss pr history for VCMLossProtectionLogic, we should take into account the last loss pr in a window
If VCMLossProtectionLogic::UpdateMaxLossHistory is called and the following condition are true:
1) now - _lossPrHistory[0].timeMs >= kLossPrShortFilterWinMs
2) this is not the first entry into _lossPrHistory
That means we are closing the current window for tracking the max loss pr and will start a new one.

As part of closing the current window we assign the current window's lossPr255 to be the current value of _shortMaxLossPr255.
_shortMaxLossPr255 should be the max loss pr we have seen in the window, however we will not take into account the lossPr255 passed in this call of VCMLossProtectionLogic::UpdateMaxLossHistory unless _shortMaxLossPr255 is 0.
We should take into account this last lossPr255 even if _shortMaxLossPr255 is not 0, as lossPr255 could very well be the largest loss pr of this window.

Bug: None
Change-Id: Iea0c8a3d43a0c5085f285a1d7994aaf94f35ffd0