Wayland screencast: always use full stream size to calculate stride
When sharing a window in GNOME, Mutter uses full desktop size for the
stream and just sets video crop metadata to let us know the real size
of the application window within the stream itself. In KDE, KWin does it
differently and the size of the stream is always equal to size of the
shared window. Our calculcated stride has to be equal to the stride of
the full desktop size. In case of KDE (KWin) this will result into the
same value, but it makes a difference in GNOME where sharing a window
and using smaller stride results into corrupted image as we would end up
copying black part of the image to our resulting frame.

Bug: chromium:1333304
Change-Id: I755a5139336c1da5abf95591a2b70a68659a255f