[ModernStats] Replace uses of std::unique_ptr<> with absl::optional<>.
Optional better describes "optionality" so let's do it for the sake of
style. But a side-effect of switching to optional may be better memory
locality than std::unique_ptr<>. Anecdotally I saw a pprof suggesting a
significant amount of time being spent allocating/reading these maps.
(This CL is unlikely to make the difference but it can't hurt.)

Bug: webrtc:14289
Change-Id: I7dcea9625b95c2f1a23e7d9595d27b58883570e2