`ApmTest.VerifyDebugDump(Int|Float)` more insightful with failures
Instead of showing individual byte differences, this CL detects
differences in the expected and actual byte streams of the evaluated
AEC dump and, if detected, parses the `audioproc::Event` proto lite
messages and calls EXPECT_EQ() for a subset of individual (sub-)fields.

Note that messages are parsed only if the byte streams of each message
pair do not match, so with no failures the test runs at no extra cost.

Bug: webrtc:7494
Change-Id: I8e0eda3b1ecfe06945b6dad5ee8871f8200d76d7